+++ Model/Volvo240.java
@@ -6,7 +6,7 @@ public class Volvo240 extends Vehicle {
 
     private final static double trimFactor = 1.25;
 
-    public Volvo240(double[] position){
+    public Volvo240(int[] position){
 
         super(4, 100, Color.black, "Model.Volvo240");
         stopEngine();
